Refactor TableComponent to inherit from AbstractTableComponent
