+++ projects/ppwcode/ng-common-components/src/lib/table/models/table-record.model.ts
@@ -1,0 +1,8 @@
+export interface TableRecord<T = unknown> {
+    /** The initial record that was passed. */
+    initialRecord: T
+    /** A local mapped representation of the record. */
+    mappedValues: Record<string, unknown>
+    /** The value generated for the trackBy function. */
+    trackByValue: unknown
+}
