+++ projects/ppwcode/ng-common-components/src/public-api.ts
@@ -12,6 +12,7 @@ export * from './lib/table/column-directives/ppw-column-header.directive'
 export * from './lib/table/empty-page/ppw-empty-table-page.directive'
 export * from './lib/table/options/table-options'
 export * from './lib/table/providers'
+export * from './lib/table/models/table-record.model'
 export * from './lib/table/table.component'
 export * from './lib/table/table.module'
 export * from './lib/loader/loader.component'
